pyproject.toml
[project]
name = "ava"
version = "0.1.0"
description = "AVA (Agile Virtual Assistant)"
authors = [
    {name = "Julien Coulombe-Morency",email = "hello@juliencm.dev"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13 <4.0"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "pydantic-ai (>=1.18.0,<2.0.0)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "botbuilder-core (>=4.17.0,<5.0.0)",
    "aiohttp (>=3.13.2,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.3"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ava"
version = "0.1.0"
packages = [{ include = "src" }]

[tool.pyright]
venvPath = "."
venv = ".venv"
